- 비트캠 제작 - 에디터에서 게임 실행 시 첫 노트 증발하는 버그 해결 중(노트 생성 시 렉이 원인으로 추정) - Metoer Stream 싱크 안 맞는 듯

+++ Assets/Scripts/OverlapEditor/TimeLineNote.cs
@@ -9,7 +9,7 @@ public class TimeLineNote : MonoBehaviour
 
     public int num;
 
-    public Dictionary<int, int> info;
+    public Dictionary<int, float> info;
 
     [HideInInspector] public Image image;
 
@@ -24,7 +24,7 @@ public class TimeLineNote : MonoBehaviour
         defaultColor = image.color;
     }
 
-    public void Setting(Dictionary<int, int> info)
+    public void Setting(Dictionary<int, float> info)
     {
         this.info = info;
     }
